+++ src/main/java/org/mule/tools/maven/plugin/module/analyze/impl/dependency/asm/DependencyClassFileVisitor.java
@@ -4,11 +4,11 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.tools.maven.plugin.module.analyze.impl.common.dependency.asm;
+package org.mule.tools.maven.plugin.module.analyze.impl.dependency.asm;
 
-import static org.mule.tools.maven.plugin.module.analyze.impl.mms.DefaultMuleModuleSystemApiAnalyzer.getPackageName;
+import static org.mule.tools.maven.plugin.module.analyze.impl.module.mms.DefaultMuleModuleSystemApiAnalyzer.getPackageName;
 
-import org.mule.tools.maven.plugin.module.analyze.impl.common.dependency.ClassFileVisitor;
+import org.mule.tools.maven.plugin.module.analyze.impl.dependency.ClassFileVisitor;
 import org.mule.tools.maven.plugin.module.analyze.api.ModuleLogger;
 
 import java.io.IOException;
