+++ src/main/java/org/mule/tools/maven/plugin/module/analyze/impl/logging/VerboseAnalyzerLogger.java
@@ -4,14 +4,22 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.tools.maven.plugin.module.analyze.impl.common.logging;
+package org.mule.tools.maven.plugin.module.analyze.impl.logging;
 
 import org.mule.tools.maven.plugin.module.analyze.api.ModuleLogger;
 
-public class SilentAnalyzerLogger implements ModuleLogger {
+import org.apache.maven.plugin.logging.Log;
+
+public class VerboseAnalyzerLogger implements ModuleLogger {
+
+  private final Log logger;
+
+  public VerboseAnalyzerLogger(Log logger) {
+    this.logger = logger;
+  }
 
   @Override
   public void log(String message) {
-    // Do nothing
+    logger.info(message);
   }
 }
